requirements.txt
aiocron==1.8
asyncio==3.4.3
certifi==2023.5.7
chardet==5.1.0
coinoxr==0.0.1a0
croniter==1.3.15
future==0.18.3
idna==3.4
natsort==8.3.1
pyaes==1.6.1
pyasn1==0.5.0
pycoingecko==3.1.0
pyjokes==0.6.0
pymongo==4.3.3
python-dateutil==2.8.2
pytz==2021.1
PyYAML==6.0
requests==2.31.0
rsa==4.9
six==1.16.0
Telethon==1.28.5
typing==3.7.4.1
tzlocal==5.0.1
ujson==5.7.0
urllib3==2.0.2
pyaml-env==1.2.1
pyjokes==0.6.0
pyaml_env==1.2.1
